backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
langchain>=0.2.0
langchain-openai>=0.1.0
langgraph>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.6.0
requests>=2.31.0
aiohttp>=3.9.0
supabase>=2.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
